feat: adds mode to support ghostscript
pdfcpu creates a pdf of version 1.7 which some printers don't
understand. Ghostscript allows downconversion. This adds a mode where
the stitching uses ghostscript instead. Will probably change this to be
the main mode.

This is slightly dirty with quite a few fmt.Println statements which
need to be cleaned up at some point.